+++ src/main/java/edu/harvard/iq/dataverse/SettingsWrapper.java
@@ -98,6 +98,7 @@ public class SettingsWrapper implements java.io.Serializable {
     //External Vocabulary support
     private Map<Long, JsonObject> cachedCvocMap = null;
     private Map<Long, JsonObject> cachedCvocByTermFieldMap = null;
+    private Set<Long> cvocFieldSet;
     
     private Long zipDownloadLimit = null; 
     
@@ -804,6 +805,17 @@ public class SettingsWrapper implements java.io.Serializable {
             }
             return cachedCvocMap;
         }
+    }
+    
+    public boolean isCvocField(Long fieldId) {
+
+        if(cvocFieldSet == null) {
+            cvocFieldSet = fieldService.getCvocFieldSet();
+        }
+        if(cvocFieldSet == null) {
+         return false;
+        }
+        return cvocFieldSet.contains(fieldId);
     }
     
     public String getMetricsUrl() {