+++ src/main/java/edu/harvard/iq/dataverse/ThumbnailServiceWrapper.java
@@ -10,6 +10,7 @@ import edu.harvard.iq.dataverse.dataaccess.ImageThumbConverter;
 import edu.harvard.iq.dataverse.dataaccess.StorageIO;
 import edu.harvard.iq.dataverse.dataset.DatasetUtil;
 import edu.harvard.iq.dataverse.search.SolrSearchResult;
+import edu.harvard.iq.dataverse.util.FileUtil;
 import edu.harvard.iq.dataverse.util.SystemConfig;
 
 import java.io.IOException;
@@ -47,6 +48,19 @@ public class ThumbnailServiceWrapper implements java.io.Serializable  {
     private Map<Long, String> dvobjectThumbnailsMap = new HashMap<>();
     private Map<Long, DvObject> dvobjectViewMap = new HashMap<>();
     private Map<Long, Boolean> hasThumbMap = new HashMap<>();
+
+    public String getFileCardImageAsUrl(SolrSearchResult result) {
+        DataFile dataFile = result != null && result.getEntity() != null ? ((DataFile) result.getEntity()) : null;
+        if (dataFile == null || result.isHarvested()
+                || !isThumbnailAvailable(dataFile)
+                || dataFile.isRestricted()
+                || !dataFile.isReleased()
+                || FileUtil.isActivelyEmbargoed(dataFile)
+                || FileUtil.isRetentionExpired(dataFile)) {
+            return null;
+        }
+        return SystemConfig.getDataverseSiteUrlStatic() + "/api/access/datafile/" + dataFile.getId() + "?imageThumb=true";
+    }
 
     // it's the responsibility of the user - to make sure the search result
     // passed to this method is of the Datafile type!
@@ -208,7 +222,13 @@ public class ThumbnailServiceWrapper implements java.io.Serializable  {
     public String getDataverseCardImageAsBase64Url(SolrSearchResult result) {
         return dataverseService.getDataverseLogoThumbnailAsBase64ById(result.getEntityId());
     }
-    
+
+    // it's the responsibility of the user - to make sure the search result
+    // passed to this method is of the Dataverse type!
+    public String getDataverseCardImageAsUrl(SolrSearchResult result) {
+        return dataverseService.getDataverseLogoThumbnailAsUrl(result.getEntityId());
+    }
+
     public void resetObjectMaps() {
         dvobjectThumbnailsMap = new HashMap<>();
         dvobjectViewMap = new HashMap<>();